requirements.txt
numpy==1.17.4
clipboard==0.0.4
Pillow==7.2.0
